requirements.txt
matplotlib
keras
numpy==1.23.5
pandas==1.5.3
ultralytics
roboflow
tensorflow-macos==2.10.0; sys_platform == 'darwin' and 'ARM' in platform_version # Mac M-chips
tensorflow==2.10.0; sys_platform == 'darwin' and 'ARM' not in platform_version # Mac Intel chips
tensorflow==2.10.0; sys_platform != 'darwin' # Windows & Ubuntu Intel chips
google-cloud-bigquery
google-cloud-storage==2.14.0
google-api-core==2.8.2
googleapis-common-protos==1.56.4
protobuf==3.19.6
python-dotenv
scikit-learn
python-multipart

# API
fastapi==0.108.0
uvicorn
streamlit
streamlit-webrtc

#For images
opencv-python
pillow
